+++ examples/tests.f90
@@ -21,7 +21,7 @@
 ! this program; see the files COPYING3 and COPYING.RUNTIME respectively.
 ! If not, see <http://www.gnu.org/licenses/>.
 !
-! Contributed by Vincent MAGNIN, 02-24-2011, last modified: 2022-04-07
+! Contributed by Vincent MAGNIN, 02-24-2011, last modified: 2022-04-24
 ! ****************
 ! Automated tests
 ! ****************
@@ -550,10 +550,18 @@ program gtk_fortran_test
   integer :: errors
   integer :: file_unit
   character(len=128) :: os_string
+  type(c_ptr) :: ret
 
   print '(A)', "Testing iso_c_binding with GTK and GLib..."
 
-  call c_f_string_copy(g_get_os_info("PRETTY_NAME"//c_null_char), os_string)
+  ! That function may return NULL with some OS:
+  ret = g_get_os_info("PRETTY_NAME"//c_null_char)
+  if (c_associated(ret)) then
+    call c_f_string_copy(ret, os_string)
+  else
+    os_string = "?"
+  end if
+  
   print '(3A,I0,A1,I0,A1,I0)', "Compiled with "//compiler_version()//" on ", TRIM(os_string), &
       & ", linked to GTK ", gtk_get_major_version(),".", gtk_get_minor_version(), ".", gtk_get_micro_version()