+++ src/gtk-fortran.f90
@@ -23,7 +23,7 @@
 !------------------------------------------------------------------------------
 ! The gtk-?-fortran command prints information about gtk-fortran.
 ! Contributors: Vincent Magnin 2021-01-29
-! Last modifications: 2022-04-06
+! Last modifications: 2022-04-24
 !------------------------------------------------------------------------------
 
 module handlers_gtk_fortran
@@ -42,13 +42,20 @@ module handlers_gtk_fortran
     type(c_ptr), value, intent(in)  :: app, gdata
     character(len=128) :: name_string, os_string
     integer :: suffix_pos
+    type(c_ptr) :: ret
 
     call c_f_string_copy(g_get_prgname(), name_string)
     ! Removing the .exe suffix if the OS is Windows:
     suffix_pos = index(name_string, ".exe")
     if (suffix_pos /= 0) name_string = name_string(1:suffix_pos-1)
 
-    call c_f_string_copy(g_get_os_info("PRETTY_NAME"//c_null_char), os_string)
+    ! That function may return NULL with some OS:
+    ret = g_get_os_info("PRETTY_NAME"//c_null_char)
+    if (c_associated(ret)) then
+      call c_f_string_copy(ret, os_string)
+    else
+      os_string = "?"
+    end if
 
     print '(2A)', TRIM(name_string), " (GTK @GTK_SEMANTIC_VERSION@ and GLib @GLIB_SEMANTIC_VERSION@)"
     print '(3A,I0,A1,I0,A1,I0)', "Compiled with "//compiler_version()//" on ", TRIM(os_string), &